# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reloading Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRecord"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "204":
          description: Password reset

  # ── Reference Data ────────────────────────────────────────────
...
components:
  schemas:
...
    UserRecord:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
        notificationsEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, role, active, notificationsEnabled, createdAt]
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        notificationsEnabled:
          type: boolean
      required: [username, email, password]
    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
        notificationsEnabled:
          type: boolean
    ResetPasswordBody:
      type: object
      properties:
        newPassword:
          type: string
      required: [newPassword]
...